fix: * symbol on tab's label, display error.
Note: Really fix it this time.

Error reason:
when run saveContent method, it use `getItemFromLocal`
method to get tab's item , which will get item from this.itemList by id,
and change it's `isChanged` attribute.

Here is the bug: when I switch to anoter group, the tab item is not in
this.itemList now.

So the `getItemFromLocal` method will call `getItemFromDb` to get item,
this the item is not match the tab's item.

Well, the solution is clear now: we can get item from the tabList.

So I add a `getItemFromTabList` method.